requirements.txt
aiogram==3.13.1
telethon==1.36.0
SQLAlchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
